rename game.state.ts
