+++ server/src/dto/chart-request.dto.ts
@@ -1,5 +1,5 @@
 import { IsNotEmpty } from 'class-validator';
-import { TimeState } from 'src/states/game.state.';
+import { TimeState } from 'src/states/game.state';
 
 export type TradeVolume = {
   buyQuantity: number;
