+++ server/src/services/item.service.ts
@@ -5,7 +5,7 @@ import { Model, Types } from 'mongoose';
 import { ItemRequestDto } from 'src/dto/item-request.dto';
 
 import { Item, ItemDocument } from 'src/schemas/item.schema';
-import { GameStateProvider } from 'src/states/game.state.';
+import { GameStateProvider } from 'src/states/game.state';
 import { isPlayer } from 'src/utis/typeGuard';
 
 @Injectable()
