+++ server/src/services/trade.service.ts
@@ -7,7 +7,7 @@ import { TradeResponseDto } from 'src/dto/trade-response.dto';
 import { Player, PlayerDocument } from 'src/schemas/player.schema';
 import { Stock, StockDocument } from 'src/schemas/stock.schema';
 import { Trade, TradeDocument } from 'src/schemas/trade.schema';
-import { GameStateProvider, TimeState } from 'src/states/game.state.';
+import { GameStateProvider, TimeState } from 'src/states/game.state';
 import { isTrade } from 'src/utis/typeGuard';
 
 @Injectable()
